DelayEmbeddings.jl

DelayEmbeddings — Module

DelayEmbeddings.jl

A Julia package that provides a generic interface for performing delay coordinate embeddings, as well as cutting edge algorithms for creating optimal embeddings given some data. It can be used as a standalone package, or as part of DynamicalSystems.jl.

To install it, run import Pkg; Pkg.add("DelayEmbeddings").

All further information is provided in the documentation, which you can either find online or build locally by running the docs/make.jl file.

Overview

Note

The documentation and the code of this package is parallelizing Chapter 6 of Nonlinear Dynamics, Datseris & Parlitz, Springer 2022.

The package provides an interface to perform delay coordinates embeddings, as explained in homonymous page.

There are two approaches for estimating optimal parameters to do delay embeddings:

Separated, where one tries to find the best value for a delay time τ and then an optimal embedding dimension d.

Unified, where at the same time an optimal combination of τ, d is found.

The separated approach is something "old school", while recent scientific research has shifted almost exclusively to unified approaches. This page describes algorithms belonging to the separated approach, which is mainly done by the function optimal_separated_de.

The unified approach is discussed in the Unified optimal embedding page.